+++ query_aggregation.go
@@ -28,6 +28,15 @@ func (o *DistinctOpts) toMap() map[string]interface{} {
 }
 
 // Distinct removes duplicate elements from the sequence.
+func Distinct(arg interface{}, optArgs ...DistinctOpts) Term {
+	opts := map[string]interface{}{}
+	if len(optArgs) >= 1 {
+		opts = optArgs[0].toMap()
+	}
+	return constructRootTerm("Distinct", p.Term_DISTINCT, []interface{}{arg}, opts)
+}
+
+// Distinct removes duplicate elements from the sequence.
 func (t Term) Distinct(optArgs ...DistinctOpts) Term {
 	opts := map[string]interface{}{}
 	if len(optArgs) >= 1 {
@@ -138,6 +147,13 @@ func (t Term) Ungroup(args ...interface{}) Term {
 // Contains returns whether or not a sequence contains all the specified values,
 // or if functions are provided instead, returns whether or not a sequence
 // contains values matching all the specified functions.
+func Contains(args ...interface{}) Term {
+	return constructRootTerm("Contains", p.Term_CONTAINS, args, map[string]interface{}{})
+}
+
+// Contains returns whether or not a sequence contains all the specified values,
+// or if functions are provided instead, returns whether or not a sequence
+// contains values matching all the specified functions.
 func (t Term) Contains(args ...interface{}) Term {
 	return constructMethodTerm(t, "Contains", p.Term_CONTAINS, args, map[string]interface{}{})
 }
@@ -148,8 +164,25 @@ func (t Term) Contains(args ...interface{}) Term {
 // Count the number of elements in the sequence. With a single argument,
 // count the number of elements equal to it. If the argument is a function,
 // it is equivalent to calling filter before count.
+func Count(args ...interface{}) Term {
+	return constructRootTerm("Count", p.Term_COUNT, funcWrapArgs(args), map[string]interface{}{})
+}
+
+// Count the number of elements in the sequence. With a single argument,
+// count the number of elements equal to it. If the argument is a function,
+// it is equivalent to calling filter before count.
 func (t Term) Count(args ...interface{}) Term {
 	return constructMethodTerm(t, "Count", p.Term_COUNT, funcWrapArgs(args), map[string]interface{}{})
+}
+
+// Sum returns the sum of all the elements of a sequence. If called with a field
+// name, sums all the values of that field in the sequence, skipping elements of
+// the sequence that lack that field. If called with a function, calls that
+// function on every element of the sequence and sums the results, skipping
+// elements of the sequence where that function returns null or a non-existence
+// error.
+func Sum(args ...interface{}) Term {
+	return constructRootTerm("Sum", p.Term_SUM, funcWrapArgs(args), map[string]interface{}{})
 }
 
 // Sum returns the sum of all the elements of a sequence. If called with a field
@@ -167,6 +200,15 @@ func (t Term) Sum(args ...interface{}) Term {
 // the sequence that lack that field. If called with a function, calls that function
 // on every element of the sequence and averages the results, skipping elements of the
 // sequence where that function returns null or a non-existence error.
+func Avg(args ...interface{}) Term {
+	return constructRootTerm("Avg", p.Term_AVG, funcWrapArgs(args), map[string]interface{}{})
+}
+
+// Avg returns the average of all the elements of a sequence. If called with a field
+// name, averages all the values of that field in the sequence, skipping elements of
+// the sequence that lack that field. If called with a function, calls that function
+// on every element of the sequence and averages the results, skipping elements of the
+// sequence where that function returns null or a non-existence error.
 func (t Term) Avg(args ...interface{}) Term {
 	return constructMethodTerm(t, "Avg", p.Term_AVG, funcWrapArgs(args), map[string]interface{}{})
 }
@@ -176,8 +218,28 @@ func (t Term) Avg(args ...interface{}) Term {
 // calls that function on every element of the sequence and returns the element
 // which produced the smallest value, ignoring any elements where the function
 // returns null or produces a non-existence error.
+func Min(args ...interface{}) Term {
+	return constructRootTerm("Min", p.Term_MIN, funcWrapArgs(args), map[string]interface{}{})
+}
+
+// Min finds the minimum of a sequence. If called with a field name, finds the element
+// of that sequence with the smallest value in that field. If called with a function,
+// calls that function on every element of the sequence and returns the element
+// which produced the smallest value, ignoring any elements where the function
+// returns null or produces a non-existence error.
 func (t Term) Min(args ...interface{}) Term {
 	return constructMethodTerm(t, "Min", p.Term_MIN, funcWrapArgs(args), map[string]interface{}{})
+}
+
+// MinIndex finds the minimum of a sequence. If called with a field name, finds the element
+// of that sequence with the smallest value in that field. If called with a function,
+// calls that function on every element of the sequence and returns the element
+// which produced the smallest value, ignoring any elements where the function
+// returns null or produces a non-existence error.
+func MinIndex(index interface{}, args ...interface{}) Term {
+	return constructRootTerm("Min", p.Term_MIN, funcWrapArgs(args), map[string]interface{}{
+		"index": index,
+	})
 }
 
 // MinIndex finds the minimum of a sequence. If called with a field name, finds the element
@@ -196,8 +258,28 @@ func (t Term) MinIndex(index interface{}, args ...interface{}) Term {
 // calls that function on every element of the sequence and returns the element
 // which produced the largest value, ignoring any elements where the function
 // returns null or produces a non-existence error.
+func Max(args ...interface{}) Term {
+	return constructRootTerm("Max", p.Term_MAX, funcWrapArgs(args), map[string]interface{}{})
+}
+
+// Max finds the maximum of a sequence. If called with a field name, finds the element
+// of that sequence with the largest value in that field. If called with a function,
+// calls that function on every element of the sequence and returns the element
+// which produced the largest value, ignoring any elements where the function
+// returns null or produces a non-existence error.
 func (t Term) Max(args ...interface{}) Term {
 	return constructMethodTerm(t, "Max", p.Term_MAX, funcWrapArgs(args), map[string]interface{}{})
+}
+
+// MaxIndex finds the maximum of a sequence. If called with a field name, finds the element
+// of that sequence with the largest value in that field. If called with a function,
+// calls that function on every element of the sequence and returns the element
+// which produced the largest value, ignoring any elements where the function
+// returns null or produces a non-existence error.
+func MaxIndex(index interface{}, args ...interface{}) Term {
+	return constructRootTerm("Max", p.Term_MAX, funcWrapArgs(args), map[string]interface{}{
+		"index": index,
+	})
 }
 
 // MaxIndex finds the maximum of a sequence. If called with a field name, finds the element