+++ query_write.go
@@ -29,7 +29,7 @@ func (t Term) Insert(arg interface{}, optArgs ...InsertOpts) Term {
 type UpdateOpts struct {
 	Durability    interface{} `gorethink:"durability,omitempty"`
 	ReturnChanges interface{} `gorethink:"return_changes,omitempty"`
-	NotAtomic     interface{} `gorethink:"non_atomic,omitempty"`
+	NonAtomic     interface{} `gorethink:"non_atomic,omitempty"`
 	Conflict      interface{} `gorethink:"conflict,omitempty"`
 }
 
@@ -52,7 +52,7 @@ func (t Term) Update(arg interface{}, optArgs ...UpdateOpts) Term {
 type ReplaceOpts struct {
 	Durability    interface{} `gorethink:"durability,omitempty"`
 	ReturnChanges interface{} `gorethink:"return_changes,omitempty"`
-	NotAtomic     interface{} `gorethink:"non_atomic,omitempty"`
+	NonAtomic     interface{} `gorethink:"non_atomic,omitempty"`
 }
 
 func (o *ReplaceOpts) toMap() map[string]interface{} {