+++ CHANGELOG.md
@@ -8,8 +8,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
  - Added support for optional arguments to `r.JS()`
  - Added `NonVotingReplicaTags` optional argument to `TableCreateOpts`
- - Added `PrimaryReplicaTag` optional argument to `ReconfigureOpts`
- - Added root term `TypeOf`, previously only the method term was supported 
+ - Added root term `TypeOf`, previously only the method term was supported
+
+### Changed
+
+ - Renamed `PrimaryTag` to `PrimaryReplicaTag` in `ReconfigureOpts`
 
 ### Fixed
  - Fixed incorrectly named `Replicas` field in `TableCreateOpts`