+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
  - Added root version of `MinIndex`
  - Added root version of `Max`
  - Added root version of `MaxIndex`
+ - Added `ReadMode` to `RunOpts`
 
 ### Changed