+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Fixed
+ - Fixed incorrectly named `Replicas` field in `TableCreateOpts`
+
 ## v2.1.2 - 2016-07-22
 
 ### Added